AI-Router-AIR/pyproject.toml
[project]
name = "ai-router-air"
version = "0.1.0"
description = "Unified API for LLM, STT, and TTS"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "pydantic-settings>=2.0.0",
    "uvloop>=0.17.0",
    "httptools>=0.6.0",
    "filelock>=3.12.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.interrogate]
ignore-private = true

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "client*"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
